refactor(consume): rename process to consume
- rename process to consume
- move batch middlewares to batch package
- move one middlewares to one package

This makes API more streamlined, batch related things are in batch
package, one-related things are in one package:

  consumer.Start(context.Background(), consume.ByOne(tx, false, one.NewDeliveryLogging(zerolog.New(os.Stdout))))
