+++ consume/batch.go
@@ -1,4 +1,4 @@
-package process
+package consume
 
 import (
 	"context"
